chore(appsec): move common context to class
